third_party/blink/web_tests/platform/generic/fast/frames/sandboxed-iframe-plugins-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE ERROR: Failed to load 'data:text/plain,' as a plugin, because the frame into which the plugin is loading is sandboxed.
CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE ERROR: Failed to load '' as a plugin, because the frame into which the plugin is loading is sandboxed.
This test verifies that sandboxing of plugins works as intended. Two tests are made, each in one sandboxed and one non-sandboxed IFrame: embeds and objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS (self.embedFrame1.document.getElementById('plugin').postMessage) != undefined is true
PASS (self.embedFrame2.document.getElementById('plugin').postMessage) == undefined is true
PASS (self.objectFrame1.document.getElementById('plugin').postMessage) != undefined is true
PASS (self.objectFrame2.document.getElementById('plugin').postMessage) == undefined is true
PASS successfullyParsed is true

TEST COMPLETE

   
